Merge Docker build and push steps
Since this image was previously only built for the target deployment's architecture, it doesn't make sense to load the image into the Docker host (`--load`) as it may not be compatible. If the image ever needs to be run locally, a separate Make command should be added in to avoid pushing a multi-arch image that's never used as such.